+++ apps/client/src/features/editor/extensions/block-id.ts
@@ -183,5 +183,3 @@ export const BlockId = new Extension({
     ];
   },
 });
-
-export { BlockId as default };
